sdelal

+++ DrawFigures/DrawFigures/Shape/Implementation/ShapeStrategy/Implementation/Point.cs
@@ -2,12 +2,12 @@
 
 public class Point
 {
-    public Point(double x, double y)
+    public Point(float x, float y)
     {
         X = x;
         Y = y;
     }
 
-    public double X { get; set; }
-    public double Y { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
 }
